requirements.txt
# task1
psycopg2-binary

# task2
hazelcast-python-client==5.4.0